spring-1.2.9

+++ spring-framework/samples/petclinic/src/org/springframework/samples/petclinic/jdbc/JdbcPet.java
@@ -3,30 +3,32 @@ package org.springframework.samples.petclinic.jdbc;
 import org.springframework.samples.petclinic.Pet;
 
 /**
+ * Subclass of Pet that carries temporary id properties
+ * which are only relevant for AbstractJdbcClinic.
+ *
  * @author Juergen Hoeller
+ * @see AbstractJdbcClinic
  */
 public class JdbcPet extends Pet {
 
-	private long typeId;
+	private int typeId;
 
-	private long ownerId;
+	private int ownerId;
 
-	public void setTypeId(long typeId) {
+	public void setTypeId(int typeId) {
 		this.typeId = typeId;
 	}
 
-	public long getTypeId() {
+	public int getTypeId() {
 		return this.typeId;
 	}
 
-	public void setOwnerId(long ownerId) {
+	public void setOwnerId(int ownerId) {
 		this.ownerId = ownerId;
 	}
 
-	public long getOwnerId() {
+	public int getOwnerId() {
 		return ownerId;
 	}
-
-
 
 }
